Delete channel, database improvements, bug fixes

+++ api/views/channels.py
@@ -1,5 +1,6 @@
 import time 
 
+import pyotp
 from flask import Blueprint, request
 
 from ..database import *
@@ -54,27 +55,27 @@ class Channels:
         creation_time = str(time.time())
 
         if not (channel := db.get_entry(CHANNEL_TABLE, id)):
-            channel = Channel(id, "-", "-", "-", creation_time, 1)
+            channel = Channel(id, None, None, None, creation_time, 1)
             db.insert_entry(CHANNEL_TABLE, channel)
 
-        if not db.get_channel_stuff([user_id, channel.id], "user_channel"):
-            db.insert_entry(USER_CHANNEL_TABLE, UserChannel(user_id, id, creation_time, None, 1, 1))
+        if not (user_channel := db.get_channel_stuff([user_id, channel.id], "user_channel")):
+            user_channel = UserChannel(user_id, id, creation_time, None, 1, 1)
+            db.insert_entry(USER_CHANNEL_TABLE, user_channel)
 
         if not (friend_channel := db.get_channel_stuff([friend_id, channel.id], "user_channel")):
-            db.insert_entry(USER_CHANNEL_TABLE, UserChannel(friend_id, id, creation_time, None, 1, 1))
+            friend_channel = UserChannel(friend_id, id, creation_time, None, 1, 1)
+            db.insert_entry(USER_CHANNEL_TABLE, friend_channel)
 
-        channel = channel.__dict__
-
-        if friend_channel and friend_channel.nick:
-            channel["display_name"] = friend_channel.nick
-        elif friend.get("display_name"):
-            channel["display_name"] = friend.get("display_name")
-
-        channel["name"] = friend.get("name")
-        channel["icon"] = friend.get("avatar")
-        channel["users"] = db.get_channel_stuff(channel["id"], "users")
-
-        return ({"channel": channel}, 200)
+        return ({"channel": {
+            **channel.__dict__,
+            "name": friend.get("name", "Deleted Account"),
+            "display_name": friend_channel.nick if friend_channel.nick else friend.get("display_name", None),
+            "icon": friend.get("avatar", "generic"),
+            "notifications": user_channel.notifications,
+            "join_time": user_channel.join_time,
+            "admin": True if user_channel.admin else False,
+            "users": db.get_channel_stuff(channel.id, "users")
+        }}, 200)
 
     @channels.route("/create", methods=["POST"])
     @Decorators.manage_database
@@ -91,9 +92,10 @@ class Channels:
         
         creation_time = str(time.time())
         channel = Channel(str(Functions.create_id(float(creation_time))), name, "generic", user_id, creation_time)
+        user_channel = UserChannel(user_id, channel.id, creation_time, None, 1)
         
         db.insert_entry(CHANNEL_TABLE, channel)
-        db.insert_entry(USER_CHANNEL_TABLE, UserChannel(user_id, channel.id, creation_time, 0, 1))
+        db.insert_entry(USER_CHANNEL_TABLE, user_channel)
         
         for member_id in users:
             if not (user := db.get_entry(USER_TABLE, member_id)):
@@ -101,14 +103,35 @@ class Channels:
             
             db.insert_entry(USER_CHANNEL_TABLE, UserChannel(user.id, channel.id, creation_time))
 
-        return ({"channel": channel}, 200)
+        return ({"channel": {
+            **channel.__dict__,
+            "nick": user_channel.nick,
+            "notifications": user_channel.notifications,
+            "join_time": user_channel.join_time,
+            "admin": True if user_channel.admin else False,
+            "users": db.get_channel_stuff(channel.id, "users")
+        }}, 200)
     
 
     # DELETE
     @channels.route("/delete/<channel_id>", methods=["DELETE"])
     @Decorators.manage_database
     @Decorators.auth
-    def delete_channel(db, channel_id):
+    def delete_channel(db, user_id, channel_id):
+        if not (channel := db.get_entry(CHANNEL_TABLE, channel_id)):
+            return ({"errors": {"channel": "Channel does not exist"}}, 400)
+        
+        if channel.owner != user_id:
+            return ({"errors": {"channel": "You are not the owner of this channel"}}, 403)
+        
+        user_settings = db.get_entry(USER_SETTING_TABLE, user_id)
+        user_secrets = db.get_entry(USER_SECRET_TABLE, user_id)
+
+        if user_settings.mfa_enabled == 0 and Security.hash_passwd(request.json.get("password"), user_secrets.password.split("$")[0]) != user_secrets.password:
+            return ({"errors": {"password": "Password doesn't match"}}, 403)
+        elif user_settings.mfa_enabled == 1 and not pyotp.TOTP(user_secrets.mfa_code).verify(request.json.get("code")):
+            return ({"errors": {"code": "Invalid two-factor code"}}, 400)
+
         db.delete_entry(None, channel_id, option="channel")
         return 200
